# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["user"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["user"]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["user"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["user"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteResult"}}}}},"tags":["user"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}},"tags":["auth"]}},"/auth/profile":{"get":{"operationId":"AuthController_profile","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/company":{"post":{"operationId":"CompanyController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Company"}}}}},"tags":["company"]},"get":{"operationId":"CompanyController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Company"}}}}}},"tags":["company"]}},"/company/{id}":{"get":{"operationId":"CompanyController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Company"}}}}},"tags":["company"]},"patch":{"operationId":"CompanyController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCompanyDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Company"}}}}},"tags":["company"]},"delete":{"operationId":"CompanyController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["company"]}},"/customer":{"post":{"operationId":"CustomerController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCustomerDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}}}},"tags":["customer"]},"get":{"operationId":"CustomerController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Customer"}}}}}},"tags":["customer"]}},"/customer/{id}":{"get":{"operationId":"CustomerController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}}}},"tags":["customer"]},"patch":{"operationId":"CustomerController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCustomerDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}}}},"tags":["customer"]},"delete":{"operationId":"CustomerController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["customer"]}}},"info":{"title":"NestJS API","description":"desc","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"password":{"type":"string"},"Company":{"type":"number"},"customers":{"type":"array","items":{"type":"string"}}},"required":["id","email","password","Company","customers"]},"User":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"password":{"type":"string"},"Company":{"type":"number"},"customers":{"type":"array","items":{"type":"string"}}},"required":["id","email","password","Company","customers"]},"UpdateUserDto":{"type":"object","properties":{}},"DeleteResult":{"type":"object","properties":{}},"RegisterDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"LoginDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"access_token":{"type":"string"}},"required":["access_token"]},"CreateCompanyDto":{"type":"object","properties":{"legalname":{"type":"string"},"taxid":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zip":{"type":"string"},"siret":{"type":"string"},"ape":{"type":"string"},"legalform":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"website":{"type":"string"},"logo":{"type":"string"},"description":{"type":"string"},"industry":{"type":"string"}},"required":["legalname","taxid","address","city","state","zip","siret","ape","legalform","phone","email","website","logo","description","industry"]},"Company":{"type":"object","properties":{"legalname":{"type":"string"},"taxid":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zip":{"type":"string"},"siret":{"type":"string"},"ape":{"type":"string"},"legalform":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"website":{"type":"string"},"logo":{"type":"string"},"description":{"type":"string"},"industry":{"type":"string"}},"required":["legalname","taxid","address","city","state","zip","siret","ape","legalform","phone","email","website","logo","description","industry"]},"UpdateCompanyDto":{"type":"object","properties":{"legalname":{"type":"string"},"taxid":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"zip":{"type":"string"},"siret":{"type":"string"},"ape":{"type":"string"},"legalform":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"website":{"type":"string"},"logo":{"type":"string"},"description":{"type":"string"},"industry":{"type":"string"}}},"CreateCustomerDto":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"password":{"type":"string"},"company":{"type":"number"}},"required":["id","email","password","company"]},"Customer":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"password":{"type":"string"},"company":{"type":"number"}},"required":["id","email","password","company"]},"UpdateCustomerDto":{"type":"object","properties":{"id":{"type":"number"},"email":{"type":"string"},"password":{"type":"string"},"company":{"type":"number"}}}}}}